Update README.md
The project involves creating a simple website dedicated to organizing a daily schedule during the holy month of Ramadan. The website aims to assist users in effectively managing their time for various activities related to spiritual and physical well-being. Each section of the website corresponds to a specific activity commonly practiced during Ramadan.