Update README.md
Project Name: Schedule Ramadan

Description:
Schedule Ramadan is a comprehensive web application designed to enhance the Ramadan experience for individuals and families. With a focus on organization, productivity, and spirituality, Schedule Ramadan provides users with a centralized platform to plan, manage, and optimize their activities throughout the holy month of Ramadan.